% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\alias{chunk_int}
\title{Chunk a vector into a list of vectors of roughly equal length.}
\usage{
chunk(x, n_chunks = NULL, max_chunk_size = NULL, method = c("seq", "mod",
  "rand"))

chunk_int(n, n_chunks = NULL, max_chunk_size = NULL, method = c("seq",
  "mod", "rand"))
}
\arguments{
\item{x}{The vector or list to be chunked.}

\item{n_chunks}{The number of chunks desired.}

\item{max_chunk_size}{The maximum desired length of any chunk.}

\item{method}{The algorithm to use for chunking.
\code{"seq"} _seq_uentially chunks \code{x}.
\code{"mod"} creates chunks where original indices are alike after modulo arithmetic.
\code{"rand"} specifies that \code{x} should be uniformly and randomly distributed across chunks.}

\item{n}{The number of integers to chunk (i.e. chunk the sequence \code{1:n}).}
}
\value{
For \code{chunk}, a list with \code{x} chunked across the list entries.
        For \code{chunk_int}, a list with the values \code{1:n} chunked across the list entries.
}
\details{
Exactly one of \code{n_chunks} and \code{max_chunk_size} should be \code{NULL} (to specify both as non-\code{NULL} is an error).
}
