% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mu_or_hu_genes.R
\name{convert_mu_or_hu_genes}
\alias{convert_mu_or_hu_genes}
\title{Convert between mouse or human genes}
\usage{
convert_mu_or_hu_genes(genelist, unique = FALSE)
}
\arguments{
\item{genelist}{A character vector of mouse: c(Gene1, Gene1) or human: c(GENE1, GENE2) genes; the function will determine whether you currently have mouse or human genes based on case.}

\item{unique}{A true/false Boolean for whether you get unique genes for a 1:1 conversion}
}
\description{
A function which uses BioMart and takes a list of human or mouse genes and returns a dataframe with inital query and the other species' homologs
}
\examples{
convert_mu_or_hu_genes("DPPA3", unique = FALSE)
convert_mu_or_hu_genes("DPPA3")
convert_mu_or_hu_genes("Dppa3", unique = FALSE)
convert_mu_or_hu_genes("Dppa3")
}
\keyword{convert}
\keyword{genes}
