% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend_size.R
\name{modify_legend_size}
\alias{modify_legend_size}
\title{A Function to Quickly Alter GGPlot legend size}
\usage{
modify_legend_size(
  my_plot,
  point_size = 0.5,
  text_size = 12,
  space_legend = 0.1,
  n_col = 1
)
}
\arguments{
\item{my_plot}{a previous generated ggplot}

\item{point_size}{defaults to 0.5, controls the dot size of legend elements}

\item{text_size}{defaults to 12, controls the size of the legend text}

\item{space_legend}{defaults to 0.1, controls the line spacing of elements of the legend}

\item{n_col}{defaults to 1, controls how many columns the legend is put in}
}
\description{
Quickly change many aspects of legend, defaulting to a size decrease; if you are using aes to assign legend variables, this might behave unpredictably.
}
\examples{
plot = add_small_legend(plot)
}
\keyword{ggplot}
\keyword{legend}
