% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\docType{data}
\name{NA_}
\alias{NA_}
\alias{NA_logical_}
\alias{NA_double_}
\title{Missing NA values.}
\format{A single (explicitly-typed) NA value.}
\usage{
NA_logical_

NA_double_
}
\description{
Explicit NA values that are missing in base R.
}
\details{
R already has \code{NA_integer_} and \code{NA_character_} for explicitly-typed NA values.
         \code{NA} is logical and \code{NA_real_} is double, but these aren't intuitive, so here are some additional explicit NA values to bring code into tighter alignment with the actual underlying base types.
}
\keyword{datasets}
