% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{notin}
\alias{notin}
\alias{\%notin\%}
\alias{\%nin\%}
\alias{\%ni\%}
\title{Not-In}
\usage{
... \%notin\% NA

... \%nin\% NA

... \%ni\% NA
}
\description{
The logical negation of R's \code{\%in\%} infix operator.
}
