% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{A simple percentilization function for tidyverse pipes (or standalone use cases)}
\usage{
percentile(x)
}
\arguments{
\item{message}{calculates percentile rank on a vector}
}
\description{
This function takes a vector of int/num and calculates percentile rank
}
\examples{
df \%>\% mutate_all(percentile)
percentile(df[[1]])
}
\keyword{percentile}
\keyword{quantile}
