% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{Special-Value-Replacement}
\alias{Special-Value-Replacement}
\alias{NA2}
\alias{Inf2}
\alias{NaN2}
\title{NA, Inf, and NaN replacement.}
\usage{
NA2(x, replacement)

Inf2(x, replacement)

NaN2(x, replacement)
}
\arguments{
\item{x}{The vector in which to replace the special value.}

\item{replacement}{The replacement value.}
}
\value{
A new vector with the special value replaced by the replacement value.
}
\description{
Replace NA, Inf, and NaN values in single vectors.
}
\details{
\code{dplyr::coalesce} has similar functionality, but is meant to deal with parallel vectors (similar to many SQL-systems's \code{COALESCE} function).
         These calls are simpler and are only intended to replace values in basic vectors.
         \code{NA2} uses \code{is.na}, \code{Inf2} uses \code{is.infinite}, and \code{NaN2} uses \code{is.nan} as the predicate function.
}
