% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{sstr}
\alias{sstr}
\title{Object structure string.}
\usage{
sstr(obj, ..., .name = FALSE, .newline = FALSE)
}
\arguments{
\item{obj}{Any R object.}

\item{...}{Arguments passed on to \code{\link[utils]{str}}.}

\item{.name}{Provide obj's variable name (as passed when \code{sstr} was called)?
(Adds an additional line to the output.)}

\item{.newline}{Trail the returned string with a newline character?}
}
\value{
A string capturing the results of \code{\link[utils]{str(obj, ...)}}.
}
\description{
Provide internal _str_ucture of an R object as a returned string.
}
\seealso{
\code{\link[utils]{str}}.
}
