% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_color_value.R
\name{transition_color_value}
\alias{transition_color_value}
\title{Function for setting to a specific color}
\usage{
transition_color_value(
  df,
  break_val = 0,
  by_val = 0.01,
  min_val = NULL,
  max_val = NULL,
  low_col = "forestgreen",
  mid_col = "black",
  high_col = "magenta"
)
}
\arguments{
\item{df}{data-frame to which values are to be synchronized}

\item{break_val}{\code{0}, or another value on which to center the transition color}

\item{by_val}{step indicator, used to ensure transitions are seamless. defaults to \code{0.01}, but \code{0.1} or \code{0.001} are also suggested}

\item{min_val}{optional, to set the minimum value to which a color can be associated, defaults to \code{min(df)}}

\item{max_val}{optional, to set the maximum value to which a color can be associated, defaults to \code{min(df)}}

\item{low_col}{specification of the color to represent the low range, defaults to \code{forestgreen}}

\item{mid_col}{specification of the color to represent the 0/midpoint transition, defaults to \code{black}}

\item{high_col}{specification of the color to represent the high range, defaults to \code{magenta}}
}
\description{
commonly used to set 0 to the transition color, or set up a custom color ramp centered on a specific value. returns a \code{col_break_list} list object with coordinated breaks and colors, with \code{col_break_list[[1]]} or \code{col_break_list$breaks} being the breaks and \code{col_break_list[[2]]} or \code{col_break_list$colors} being the colors
}
\examples{
transition_color_value(df)
transition_color_value(df, break_val = 0, low_col = "blue", mid_col = "yellow", high_col = "red")
}
\seealso{
\code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}
}
