% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpercent.R
\name{unpercent}
\alias{unpercent}
\title{A simple unpercenting function for tidyverse pipes (or standalone use cases)}
\usage{
unpercent(x)
}
\arguments{
\item{message}{divides what you pass to it by 100}
}
\description{
This function takes a vector of int/num and divides it by 100
}
\examples{
df \%>\% mutate_all(unpercent)
unpercent(df[[1]])
}
\keyword{percent}
\keyword{unpercent}
