% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R
\docType{data}
\name{zzz-ggproto}
\alias{zzz-ggproto}
\alias{GeomSplitViolin}
\title{ggproto Classes for zzz}
\format{
An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSplitViolin
}
\description{
ggproto Classes for zzz
}
\seealso{
\code{\link[ggplot2:ggproto]{ggplot2::ggproto()}}
}
\keyword{datasets}
